+++ docs/conf.py
@@ -20,8 +20,8 @@
 project = 'FeatherDoc'
 copyright = '2026, wuxianggujun and upstream contributors'
 author = 'wuxianggujun and upstream contributors'
-version = '1.1.0'
-release = '1.1.0'
+version = '1.2.0'
+release = '1.2.0'
 
 # -- General configuration ---------------------------------------------------